update README + CLAUDE.md for new project structure, fix DMS AMP dtype
- Rewrite README.md: 6 pruning methods, 4 KD methods, benchmark table,
  updated paths for reorganized folder structure
- Fix DMS soft mask dtype: cast mask to output.dtype in hook instead of
  removing/re-registering hooks around validation
- Add tools/plot_prune_benchmark.py for visualizing pruning comparison

+++ dms/dms_utils.py
@@ -272,8 +272,8 @@ def make_soft_mask_hook(bn_name, a_params, importance='gamma', taylor_buffers=No
                         )
             mask.register_hook(_taylor_backward_hook)
 
-        # Step 5: Apply mask
-        return output * mask.view(1, -1, 1, 1)
+        # Step 5: Apply mask (cast to output dtype to avoid float32/float16 mismatch with AMP)
+        return output * mask.to(dtype=output.dtype).view(1, -1, 1, 1)
 
     return hook
